requirements.txt
cffi==1.15.0
cryptography==35.0.0
javaobj-py3==0.4.3
pyasn1==0.4.8
pyasn1_modules==0.2.8
pycparser==2.21
pycryptodome==3.11.0
pycryptodomex==3.11.0
pyjks==20.0.0
twofish==0.3.0
